Allow deep cloning of Objects to preserve state in CT tests.
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@554 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ core/vdmj/src/main/java/org/overturetool/vdmj/traces/RepeatTraceNode.java
@@ -54,20 +54,20 @@ public class RepeatTraceNode extends TraceNode
 
 		for (int r = from; r <= to; r++)
 		{
- 			int[] c = new int[(int)r];
-			
+ 			int[] c = new int[r];
+
 			for (int i=0; i<r; i++)
 			{
 				c[i] = count;
 			}
-			
+
 			Permutor p = new Permutor(c);
-			
+
 			while (p.hasNext())
 			{
 	   			CallSequence seq = new CallSequence();
 	   			int[] select = p.next();
-	   			
+
 	   			for (int i=0; i<r; i++)
     			{
     				seq.addAll(rtests.get(select[i]));
